chrome-extension/requirements.txt
streamlit
newspaper
nltk
sumy.parsers.plaintext
sumy.nlp.tokenizers
sumy.summarizers.lsa